Refactor login code to new page
